Bael 3653: Adding Pageable Support (#8766)
* Removed the commented code

* Upgrading the Spring Doc Version

* Adding the Pageable Example

In order to use `Pageable`, I had to add `spring-data-jpa` dep and H2. Also, the Spring doc version was quite outdated, so an upgrade for that, too.

* Adding the Pageable Example

In order to use `Pageable`, I had to add `spring-data-jpa` dep and H2. Also, the Spring doc version was quite outdated, so an upgrade for that, too.

Former-commit-id: 5f5e4c0c41607ac066d93dda35c2e0c6e90aae34